moved _xml_util.py back into junitxml.py
